+++ libs/opengl/include/mrpt/opengl/CCamera.h
@@ -9,30 +9,41 @@
 
 #pragma once
 
+#include <mrpt/img/TCamera.h>
 #include <mrpt/opengl/CRenderizable.h>
 #include <mrpt/poses/CPoseOrPoint.h>
 
+#include <optional>
+
 namespace mrpt::opengl
 {
-class COpenGLViewport;
-class CCamera;
-
-/** A camera: if added to a scene, the viewpoint defined by this camera will be
- * used instead of the camera parameters set in COpenGLViewport::m_camera.
- *  A camera can be defined to provide a projective or orthogonal view of the
- * world by setting the member CCamera::m_projectiveModel.
+/** Defines the intrinsic and extrinsic camera coordinates from which to render
+ * a 3D scene.
  *
- * Alternatively, a camera pose and orientation can be given as a CPose3D object
- * if set6DOFMode() is set to `true`, then the pose can be changed with
- * `CRenderizable::setPose()`. Pose axis convention is +X pointing forwards, +Z
- * up.
+ *  By default, each viewport has its own camera, accesible via
+ *  COpenGLViewport::getCamera(), but if a CCamera object is added as an object
+ *  to be rendered, it will override the internal viewport camera.
+ *
+ *  Projection models:
+ *  - Projective model, parameterized via setProjectiveFOVdeg() (vertical field
+ * of view, in degrees), or
+ *  - Projective model, by means of a computer vision pinhole intrinsic
+ * parameter set (see `setProjectiveFromPinhole()`), or
+ *  - Orthogonal projection model (use `setProjectiveModel(false)`, or
+ * `setOrthogonal()`).
+ *
+ * Placing cameras can be done:
+ * - Using an "orbit" model: defined by a "pointing to" point, a distance to
+ * object, and azimuth + elevation angles; or
+ * - Directly giving the SE(3) camera pose, setting the set6DOFMode() to `true`
+ * and storing the desired pose with CRenderizable::setPose(). Pose axis
+ * convention is +Z pointing forwards, +X right, +Y down.
  *
  *  \sa opengl::COpenGLScene
  * \ingroup mrpt_opengl_grp
  */
 class CCamera : public CRenderizable
 {
-	friend class COpenGLViewport;
 	DEFINE_SERIALIZABLE(CCamera, mrpt::opengl)
    public:
 	CCamera() = default;
@@ -43,6 +54,44 @@ class CCamera : public CRenderizable
 		// None: a camera is a non-visual element.
 		return {};
 	}
+
+	/** @name Projection model (camera intrinsic parameters)
+	 *  @{ */
+
+	/** Enable/Disable projective mode (vs. orthogonal). */
+	void setProjectiveModel(bool v = true) { m_projectiveModel = v; }
+
+	/** Enable/Disable orthogonal mode (vs. projective)*/
+	void setOrthogonal(bool v = true) { m_projectiveModel = !v; }
+
+	void setProjectiveFromPinhole(const mrpt::img::TCamera& camIntrinsics)
+	{
+		m_projectiveModel = true;
+		m_pinholeModel = camIntrinsics;
+	}
+
+	/** Vertical field-of-View in degs, only when projectiveModel=true
+	 * (default=30 deg).
+	 */
+	void setProjectiveFOVdeg(float ang) { m_projectiveFOVdeg = ang; }
+	/** Field-of-View in degs, only when projectiveModel=true (default=30 deg).
+	 */
+	float getProjectiveFOVdeg() const { return m_projectiveFOVdeg; }
+
+	bool hasPinholeModel() const { return m_pinholeModel.has_value(); }
+
+	const std::optional<mrpt::img::TCamera>& getPinholeModel() const
+	{
+		return m_pinholeModel;
+	}
+
+	bool isProjective() const { return m_projectiveModel; }
+	bool isOrthogonal() const { return !m_projectiveModel; }
+
+	/** @} */
+
+	/** @name Defines camera pose (camera extrinsic parameters)
+	 *  @{ */
 
 	void setPointingAt(float x, float y, float z)
 	{
@@ -66,34 +115,29 @@ class CCamera : public CRenderizable
 	float getPointingAtX() const { return m_pointingX; }
 	float getPointingAtY() const { return m_pointingY; }
 	float getPointingAtZ() const { return m_pointingZ; }
+
+	mrpt::math::TPoint3Df getPointingAt() const
+	{
+		return {m_pointingX, m_pointingY, m_pointingZ};
+	}
+
 	void setZoomDistance(float z) { m_eyeDistance = z; }
 	float getZoomDistance() const { return m_eyeDistance; }
 	float getAzimuthDegrees() const { return m_azimuthDeg; }
 	float getElevationDegrees() const { return m_elevationDeg; }
 	void setAzimuthDegrees(float ang) { m_azimuthDeg = ang; }
 	void setElevationDegrees(float ang) { m_elevationDeg = ang; }
-	/** Enable/Disable projective mode (vs. orthogonal) */
-	void setProjectiveModel(bool v = true) { m_projectiveModel = v; }
-	/** Enable/Disable orthogonal mode (vs. projective) */
-	void setOrthogonal(bool v = true) { m_projectiveModel = !v; }
-	/** Set 6DOFMode, if enabled camera is set according to its pose
-	 *(default=false).
-	 *  Conventionally, eye is set looking towards the positive direction of Z
-	 *axis.
-	 *  Up is set as the Y axis.
-	 *  In this mode azimuth/elevation are ignored.
-	 **/
+
+	/** Set 6DOFMode, if enabled camera is set according to its pose, set via
+	 *CRenderizable::setPose(). (default=false).
+	 * Conventionally, eye is set looking towards +Z axis, "down" is the +Y
+	 * axis, right is "+X" axis. In this mode azimuth/elevation are ignored.
+	 */
 	void set6DOFMode(bool v) { m_6DOFMode = v; }
-	bool isProjective() const { return m_projectiveModel; }
-	bool isOrthogonal() const { return !m_projectiveModel; }
+
 	bool is6DOFMode() const { return m_6DOFMode; }
-	/** Vertical field-of-View in degs, only when projectiveModel=true
-	 * (default=30 deg).
-	 */
-	void setProjectiveFOVdeg(float ang) { m_projectiveFOVdeg = ang; }
-	/** Field-of-View in degs, only when projectiveModel=true (default=30 deg).
-	 */
-	float getProjectiveFOVdeg() const { return m_projectiveFOVdeg; }
+
+	/** @} */
 
 	/** Render does nothing here. */
 	void render(const RenderContext& rc) const override {}
@@ -114,9 +158,15 @@ class CCamera : public CRenderizable
 	/** If set to true (default), camera model is projective, otherwise, it's
 	 * orthogonal. */
 	bool m_projectiveModel{true};
-	/** Field-of-View in degs, only when projectiveModel=true (default=30 deg).
+
+	/** If defined, it overrides m_projectiveFOVdeg. */
+	std::optional<mrpt::img::TCamera> m_pinholeModel;
+
+	/** Field-of-View in degs, only when projectiveModel=true and there is no
+	 * pinhole model assigned. (default=30 deg).
 	 */
 	float m_projectiveFOVdeg{30};
+
 	/** If set to true, camera pose is used when rendering the viewport */
 	bool m_6DOFMode{false};
 };