+++ libs/opengl/src/CCamera.cpp
@@ -12,6 +12,7 @@
 #include <mrpt/opengl/CCamera.h>
 #include <mrpt/opengl/opengl_api.h>
 #include <mrpt/serialization/CArchive.h>
+#include <mrpt/serialization/optional_serialization.h>
 
 using namespace mrpt;
 using namespace mrpt::opengl;
@@ -20,13 +21,14 @@ using namespace std;
 
 IMPLEMENTS_SERIALIZABLE(CCamera, CRenderizable, mrpt::opengl)
 
-uint8_t CCamera::serializeGetVersion() const { return 1; }
+uint8_t CCamera::serializeGetVersion() const { return 2; }
 void CCamera::serializeTo(mrpt::serialization::CArchive& out) const
 {
 	// Save data:
 	out << m_pointingX << m_pointingY << m_pointingZ << m_eyeDistance
 		<< m_azimuthDeg << m_elevationDeg << m_projectiveModel
 		<< m_projectiveFOVdeg;
+	out << m_pinholeModel;	// v2
 }
 
 void CCamera::serializeFrom(mrpt::serialization::CArchive& in, uint8_t version)
@@ -34,11 +36,15 @@ void CCamera::serializeFrom(mrpt::serialization::CArchive& in, uint8_t version)
 	switch (version)
 	{
 		case 1:
+		case 2:
 		{
 			// Load data:
 			in >> m_pointingX >> m_pointingY >> m_pointingZ >> m_eyeDistance >>
 				m_azimuthDeg >> m_elevationDeg >> m_projectiveModel >>
 				m_projectiveFOVdeg;
+			if (version >= 2) in >> m_pinholeModel;
+			else
+				m_pinholeModel.reset();
 		}
 		break;
 		case 0:
@@ -54,10 +60,5 @@ void CCamera::serializeFrom(mrpt::serialization::CArchive& in, uint8_t version)
 /** In this class, returns a fixed box (max,max,max), (-max,-max,-max). */
 auto CCamera::getBoundingBox() const -> mrpt::math::TBoundingBox
 {
-	return {
-		{std::numeric_limits<double>::max(), std::numeric_limits<double>::max(),
-		 std::numeric_limits<double>::max()},
-		{-std::numeric_limits<double>::max(),
-		 -std::numeric_limits<double>::max(),
-		 -std::numeric_limits<double>::max()}};
+	return mrpt::math::TBoundingBox::PlusMinusInfinity();
 }